test: decompose oversized rust test files
